fix/rename extra component name
