+++ src/features/tutor-application/ui/diplomas-upload/index.module.scss
