+++ src/features/tutor-application/ui/index.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 
-import Button from '../../shared/ui/button';
-import Stepper from '../../shared/ui/stepper';
-import CertificatesBlock from './certificatesBlock/ui';
+import Button from '../../../shared/ui/button';
+import Stepper from '../../../shared/ui/stepper';
+import DiplomasUpload from './diplomas-upload';
 
 import styles from './index.module.scss';
 
-const TutorQuestionnaire: React.FC = () => {
+const TutorApplication: React.FC = () => {
   return (
     <div className={styles.container}>
       <div className={styles.questionnaire}>
         <h2 className={styles.questionnaire__title}>Анкета</h2>
         <Stepper currentStep={3} totalSteps={5} />
-        <CertificatesBlock />
+        <DiplomasUpload />
       </div>
       <div className={styles.container__buttons}>
         <Button
@@ -32,4 +32,4 @@ const TutorQuestionnaire: React.FC = () => {
   );
 };
 
-export default TutorQuestionnaire;
+export default TutorApplication;
